+++ ProjectSettings/EditorBuildSettings.asset
@@ -5,10 +5,13 @@ EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
   m_Scenes:
-  - enabled: 1
+  - enabled: 0
     path: Assets/Scenes/MainMenu.unity
     guid: e29e949efdc9bf1499224e88d9edb5cb
+  - enabled: 0
+    path: Assets/Scenes/MainGameZhenya.unity
+    guid: e4ffd2f38ca844a47b3bdbc168993b87
   - enabled: 1
     path: Assets/Scenes/MainGameValya.unity
-    guid: e4ffd2f38ca844a47b3bdbc168993b87
+    guid: e4e5b95a2592e3b4083ae48a66bc8c17
   m_configObjects: {}
